+++ src/test/java/tfw/immutable/ila/floatila/FloatIlaFromCastIntIlaTest.java
@@ -24,7 +24,7 @@ class FloatIlaFromCastIntIlaTest {
         IntIla ila = IntIlaFromArray.create(array);
         FloatIla targetIla = FloatIlaFromArray.create(target);
         FloatIla actualIla = FloatIlaFromCastIntIla.create(ila);
-        final float epsilon = (float) 0.0;
+        final float epsilon = 0.0f;
         FloatIlaCheck.checkAll(
                 targetIla,
                 actualIla,
